daily-status-project/user-login-register/requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
Werkzeug==2.2.2
